requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
idna==3.4
requests==2.31.0
sejong-univ-auth==0.3.4
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
djangorestframework==3.14.0